+++ spec/dummy/client/entries/app-kitten.js
@@ -31,7 +31,7 @@ import SelectWithState from 'kitten/components/form/select-with-state'
 import Slider from 'kitten/components/form/slider'
 import Switch from 'kitten/components/form/switch'
 import { TextInput } from 'kitten/components/form/text-input'
-import { TextInputWithCounter } from 'kitten/components/form/text-input-with-counter'
+import { TextInputWithLimit } from 'kitten/components/form/text-input-with-limit'
 
 // Grid
 import { Grid, GridCol } from 'kitten/components/grid/grid'
@@ -160,7 +160,7 @@ ReactOnRails.register({
   Slider,
   Switch,
   TextInput,
-  TextInputWithCounter,
+  TextInputWithLimit,
 
   // Icons
   GrabberIcon,