+++ spec/dummy/client/javascripts/karl/components/popovers/auto-trigger-popover.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { autoTriggerEnhancer } from 'kitten/hoc/auto-trigger-enhancer'
+import { triggerEnhancer } from 'kitten/hoc/trigger-enhancer'
 import { Popover } from 'kitten/components/popovers/popover'
 
 const text = 'I could pee on this if i had the energy. Present belly, scratch ' +
@@ -7,7 +7,7 @@ const text = 'I could pee on this if i had the energy. Present belly, scratch ' 
              'of string. Stare at ceiling light why must they do that, for chase' +
              'laser chase imaginary bugs.'
 
-export const KarlAutoTriggerPopover = autoTriggerEnhancer(Popover, {
+export const KarlAutoTriggerPopover = triggerEnhancer(Popover, {
   children: text,
   popoverClassName: 'karl-AutoTriggerPopover',
   titleAriaLabelId: 'dialogtitle-auto-trigger-popover',